--- packages/localstorage-cache/src/cache/index.ts
@@ -1,4 +1,0 @@
-const version = '__@FLOPFLIP/VERSION_OF_RELEASE__';
-
-export { version };
-export { createCache } from './cache';
